+++ src/main/java/dinkplugin/message/NotificationBody.java
@@ -72,6 +72,22 @@ public class NotificationBody<T extends NotificationData> {
     String computedDiscordContent;
 
     /**
+     * An optional title to override that of {@link NotificationType#getThumbnail}
+     * within the embed constructed by {@link DiscordMessageHandler#createMessage}
+     */
+    @Nullable
+    @EqualsAndHashCode.Include
+    transient String customTitle;
+
+    /**
+     * An optional footer text to override that of {@link DinkPluginConfig#embedFooterText}
+     * within the embed constructed by {@link DiscordMessageHandler#createMessage}
+     */
+    @Nullable
+    @EqualsAndHashCode.Include
+    transient String customFooter;
+
+    /**
      * An optional thumbnail to override that of {@link NotificationType#getThumbnail}
      * within the embed constructed by {@link DiscordMessageHandler#createMessage}
      */